feat(UserInput): add 금액 입력 및 검증 함수

+++ src/main/java/lotto/view/output/Guide.java
@@ -5,7 +5,7 @@ public class Guide {
         System.out.println("구입금액을 입력해 주세요.");
     }
 
-    public void printLottoCount(int lottoCount) {
+    public void printLottoCount(long lottoCount) {
         System.out.printf("%d개를 구매했습니다.%n", lottoCount);
     }
